+++ resources/views/core/layout/menu.blade.php
@@ -25,6 +25,9 @@
             <a class="admin-nav-link" href="{{ route('admin.user.index') }}">Users</a>
         </li>
         <li class="mr-3">
+            <a class="admin-nav-link" href="{{ route('admin.view.index') }}">Views</a>
+        </li>
+        <li class="mr-3">
             <a class="admin-nav-link" href="{{ route('feeds.main') }}" target="_blank">RSS-feed</a>
         </li>
         <li class="mr-3">